update manage groups dialog location and code
