+++ src/app/components/manage-groups-dialog.component.html
