+++ src/app/dashboard/services/dashboard-index.service.ts
@@ -1,6 +1,4 @@
-import { TaskStatus } from '@aneoconsultingfr/armonik.api.angular';
 import { Injectable, inject } from '@angular/core';
-import { TasksStatusesService } from '@app/tasks/services/tasks-statuses.service';
 import { DefaultConfigService } from '@services/default-config.service';
 import { DashboardStorageService } from './dashboard-storage.service';
 import { Line } from '../types';
@@ -9,29 +7,9 @@ import { Line } from '../types';
 export class DashboardIndexService {
   #defaultConfigService = inject(DefaultConfigService);
   #dashboardStorageService = inject(DashboardStorageService);
-  #tasksStatusesService = inject(TasksStatusesService);
 
   readonly defaultLines: Line[] = this.#defaultConfigService.defaultDashboardLines;
   readonly defaultSplitLines: number = this.#defaultConfigService.defaultDashboardSplitLines;
-
-  // TODO: move to TasksStatusesService
-  statuses(): { value: string, name: string }[] {
-    const values = Object.values(this.#tasksStatusesService.statuses).sort((a, b) => a.toString().localeCompare(b.toString()));
-    const keys = Object.keys(this.#tasksStatusesService.statuses).sort((a, b) => a.toString().localeCompare(b.toString()));
-    const sortedKeys = values.map((value) => {
-      return keys.find((key) => {
-        return this.#tasksStatusesService.statuses[Number(key) as TaskStatus] === value;
-      });
-    });
-
-    return (sortedKeys.filter(Boolean) as string[]).map((key) => {
-      const status = Number(key) as TaskStatus;
-      return {
-        value: key,
-        name: this.#tasksStatusesService.statusToLabel(status)
-      };
-    });
-  }
 
   addLine(line: Line): number | void {
     const lines = this.restoreLines();
